moved the views into a separated module
