+++ estimator/views/__init__.py
@@ -1,0 +1,9 @@
+from .archest_activity_viewset import ActivityViewSet
+from .archest_activity_work_entries_viewset import ActivityWorkEntriesViewSet
+from .archest_customer_viewset import CustomerViewSet
+from .archest_estimate_viewset import EstimateViewSet
+from .archest_phase_viewset import PhaseViewSet
+from .archest_project_viewset import ProjectViewSet
+from .archest_resource_viewset import ResourceViewSet
+from .archest_sub_activity_viewset import SubActivityViewSet
+from .archest_sub_activity_work_entries_viewset import SubActivityWorkEntriesViewSet
